+++ SLUGS/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "2.3.0IP"
+__version__ = "2.3.0"
 VERSION = __version__  # synonym